pyproject.toml
[project]
name = "github-sync-fork"
version = "0.1.0"
description = "A simple project to programmatically do all github repository sync in Python."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.11",
    "pygithub>=2.6.1",
    "pytest>=8.4.0",
    "requests>=2.32.3",
    "ruff>=0.11.13",
    "click>=8.0.0",
]

[project.scripts]
github-sync-fork = "main:sync_all_forks_cli"
